LayoutTests/imported/w3c/web-platform-tests/screen-orientation/lock-sandboxed-iframe-expected.txt



Harness Error (TIMEOUT), message = null

FAIL Test without 'allow-orientation-lock' sandboxing directive assert_equals: screen.lockOrientation() throws a SecurityError expected "SecurityError" but got "NotSupportedError"
TIMEOUT Test with 'allow-orientation-lock' sandboxing directive Test timed out
